--- a/Minio/Minio.csproj
+++ b/Minio/Minio.csproj
@@ -31,6 +31,8 @@
     <PackageReference Include="AsyncFixer" Version="1.1.6" PrivateAssets="All" />
 
     <PackageReference Include="RestSharp" Version="106.3.1" />
+
+    <PackageReference Include="System.Net.Http" Version="4.3.4" />
     <PackageReference Include="System.Reactive.Linq" Version="4.0.0" />
   </ItemGroup>