Hi!

I found that you using streams wrong
It's very bad with real big files

![image](https://user-images.githubusercontent.com/5781038/47038159-38a05300-d19a-11e8-9125-62c370ca2ccf.png)

Below performance tests and PR only for review good approaches
I didn't know how to use RestSharp, but it's enough for demo

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Core i7-4770 CPU 3.40GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.402
  [Host]     : .NET Core 2.1.4 (CoreCLR 4.6.26814.03, CoreFX 4.6.26814.02), 64bit RyuJIT
  Job-KQRNGT : .NET Core 2.1.4 (CoreCLR 4.6.26814.03, CoreFX 4.6.26814.02), 64bit RyuJIT

Server=False  

```
|                   Method |    Mean |    Error |   StdDev |     Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|------------------------- |--------:|---------:|---------:|----------:|----------:|----------:|-----------:|
|     PutObjectAsync_small | 2.156 s | 0.0416 s | 0.0408 s | 1000.0000 | 1000.0000 | 1000.0000 | 3137.93 KB |
| PutObjectAsyncFast_small | 2.137 s | 0.0413 s | 0.0507 s |         - |         - |         - |   60.06 KB |
|       PutObjectAsync_big | 7.121 s | 0.0878 s | 0.0779 s | 1000.0000 | 1000.0000 | 1000.0000 | 6221.46 KB |
|   PutObjectAsyncFast_big | 2.847 s | 0.0553 s | 0.0637 s |         - |         - |         - |   60.06 KB |

[BenchmarkDotNet.Artifacts.zip](https://github.com/minio/minio-dotnet/files/2484446/BenchmarkDotNet.Artifacts.zip)