@poornas 2018-10-17
@SanSYS , thanks for the PR. minio-dotnet will need to implement streaming chunked uploads to improve performance on large files. we will try to add this in the near future.