@poornas 2018-10-17
@SanSYS, please feel free to send a PR. streaming chunked signature needs to be implemented similar to https://github.com/minio/minio-go sdk